requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
pathlib